sync: implement futex-based Mutex
This is largely based on this paper by Ulrich Drepper:
https://dept-info.labri.fr/~denis/Enseignement/2008-IR/Articles/01-futex.pdf

See the comment in the source code for details.
